+++ src/main/scala/com/signalcollect/approx/flood/WRMI.scala
@@ -39,9 +39,9 @@ import com.signalcollect.interfaces.MessageBus
  *  @param constraints: the set of constraints in which it is involved
  *  @param possibleValues: which values can the state take
  */
-class WRMIVertex(id: Any, constraints: Iterable[Constraint], possibleValues: Array[Int]) extends DataGraphVertex(id, 0.0) {
+class WRMIVertex(id: Any, constraints: Iterable[Constraint], possibleValues: Array[Int]) extends DataGraphVertex(id, 0) {
 
-  type Signal = Double
+  type Signal = Int
   var inertia: Double = 0.5 //time counter used for calculating temperature
   var oldState = possibleValues(0)
   var weightedAvgDiff: Array[Double] = Array.fill[Double](possibleValues.size)(0)
@@ -53,20 +53,20 @@ class WRMIVertex(id: Any, constraints: Iterable[Constraint], possibleValues: Arr
    * The collect function chooses a new random state and chooses it if it improves over the old state,
    * or, if it doesn't it still chooses it (for exploring purposes) with probability decreasing with time
    */
-  def collect(oldState: Double, mostRecentSignals: Iterable[Double]): Double = {
+  def collect(oldState: Int, mostRecentSignals: Iterable[Int]): Int = {
 
     //Update the weighted average regrets for each action 
     val neighbourConfigs = mostRecentSignalMap.map(x => (x._1,x._2)).toMap //neighbourConfigs must be immutable and mostRecentSignalMap is mutable, so we convert
 
     //we calculate the utility for our last state
-    val configs = neighbourConfigs + (id -> oldState.toDouble)
+    val configs = neighbourConfigs + (id -> oldState)
     val utility = constraints.foldLeft(0.0)((a, b) => a + b.utility(configs))
 
     var normFactor: Double = 0
 
     for (i <- 0 to (possibleValues.size - 1)) {
       state = possibleValues(i)
-      val possibleStatesConfigs = neighbourConfigs + (id -> state.toDouble)
+      val possibleStatesConfigs = neighbourConfigs + (id -> state)
       val possibleStatesConfigsUtility = constraints.foldLeft(0.0)((a, b) => a + b.utility(possibleStatesConfigs))
       val regret = possibleStatesConfigsUtility - utility
       weightedAvgDiff(i) = fadingMemory * regret + (1 - fadingMemory) * weightedAvgDiff(i)
@@ -96,7 +96,7 @@ class WRMIVertex(id: Any, constraints: Iterable[Constraint], possibleValues: Arr
     }
 
     val candidateState = possibleValues(candidateStateIndex)
-    val candidateStateConfigs = neighbourConfigs + (id -> candidateState.toDouble)
+    val candidateStateConfigs = neighbourConfigs + (id -> candidateState)
 
     val probabilityAcc: Double = r.nextDouble()
